create delete post function
